Move BackOff tests to correct package
